More caching fixes.

+++ src/Cassette.UnitTests/TempDirectory.cs
@@ -20,7 +20,10 @@ namespace Cassette
 
         public void Dispose()
         {
-            Directory.Delete(path, true);
+            if (Directory.Exists(path))
+            {
+                Directory.Delete(path, true);
+            }
         }
     }
 }
